Add published field to events (#135)
* Add published field to events

This commit adds a published field to events which determines whether or not events should be published
in the apps and on the website. Previously there was a public_registration field, but that field is supposed
to be used in the future for event authentication purposes.

* Update test for events api

* Update admin view for Events

* Hide public registration field in the admin

It's not currently being used anywhere.