Hid matching fields in the CR
The fields used to collect data for the student matching in the CR are
not yet fully implemented, but we need to get the CR up and running
anyway. Therefore, those fields are now temporarily hidden until they
have been properly implemented. Companies can finish the CR anyway, but
we'll have to ask them to return and fill our the matching fields once
they're up.