Merging of company registration form into the user registration form
The form in /register/company has been 'fused' with the one in /register/user

Before, if users could not find their company in /register/user, they had to click on a link which would bring them to /register/company. There, they could fill a form (very similar to the one in /register/user) in order to register their user and their company into our DB.

Since both forms are relatively similar, this fusion has been done so that users find everything more handy, with the goal of making the registration process easier for them.

Now, if the company is not displayed in the browser, users can fill their company information in /register/user as an alternative to selecting a company from the browser.