feat: CRM pagination (#855)
## Describe your changes

Fixes: #847 

Adds pagination to the CRM page. To do this, the filter form had to be
changed from post to get to give URL parameters, and filtering had to be
changed to SQL level.

A constant (COMPANIES_PER_PAGE) controls how many companies are viewed
per page. Perhaps this could be made into a UI function (maybe even part
of the form) so that the user can select how many companies they want
per page.

The free text search has been reworked to work with the pagination. It
ranks results based on simple relevance (i.e search query matches on
company names are ranked higher than matches on other fields).


To be added:

    - More ranking by relevance?
    
- (if needed) implement search engine like searching tools (such as
prepending "group:" to a query to run the search on group names etc)


## Checklist before requesting review

- [x] Feature/fix PRs should add one atomic (as small as possible)
feature or fix.
- [x] The code compiles and all the tests pass.

---------

Co-authored-by: eliasmar <eliasmar@kth.se>