Add waiting list registering (#1286)
## Waiting List

(waiting list is only for free events, since no deregistering is allowed
in payed events)

Note: if an event is fully booked and people have started putting
themselves on the waitlist, raising the maximum number of allowed
participants will not give the places to the people in the waiting
lists. Instead, these spots will simply be open to everyone else! (Not
ideal...)

- [x] add waiting list button
- [x] let people sign up to the waiting list
- [x] add waiting list tab in Events
- [x] when someone deregisters, remove first person from waiting list
(in order of sign-up) and change their status to full participant
- [x] send them an email to inform them they have gotten a spot

![bild](https://github.com/user-attachments/assets/0293708d-3fbf-4058-a63a-2a99737a8382)

![bild](https://github.com/user-attachments/assets/336886e0-2b3b-4b00-9288-09276ff2cc0c)

![bild](https://github.com/user-attachments/assets/77dfe82b-5183-42b5-bcf5-fe8f78179a16)

![bild](https://github.com/user-attachments/assets/7e938c3d-d409-47d6-b253-e5f2b2b4c30d)

Things that need to be fixed:
- [ ] visualize number of waiting list participants in events tab
- [x] every instance where participants is called, filter away the
people in waiting list (if needed)
- [x] when sign ups are closed but waiting list signup is enabled, make
it possible to actually sign up to the waiting list
- [x] make it possible to view waiting list participants in admin ais
website

---------

Co-authored-by: Lucia Karens <Lucia Karens>
Co-authored-by: Didrik Munther <dmu0817@gmail.com>